Update README.md
This project predicts house prices using a linear regression model based on key features like square footage, number of bedrooms, and bathrooms. Publicly available datasets from Kaggle were used for training and evaluation. The data was cleaned and visualized to understand feature relationships. The model was evaluated using MAE, MSE, RMSE, and R² score, and tested on custom sample inputs to simulate real-world use. This is a basic, beginner-friendly implementation and serves as a foundation for future improvements